Update v1 of driver:
Were added :

- voltage and current limits functions (voltage_limit_up, voltage_limit_low, current_limit_up, current_limit_low; current & voltage limits do not apply to BE2141)
- activating automatic range selections (curr_range_auto for BE5845 only, voltage_range_auto for BE2141/BE2142)
- current range selection for BE5845
- limits monitoring (activating them or not: parameter "limits_monitoring"; not possible to desactivate on BE5845)

Now the number of the channels is automatically detected from the physical number in the chassis so that the 'num_channels' argument is not anymore needed when calling the class in a notebook.
The name of the channels is as follow : 'modX_chYY' where X is the position of the module in the chassis (from 1 to 5) and YY is the channel (output) number in the considered module (e.g. YY = 02 if 2nd output of the board).

Added some various securities and error  messages when input mistakes to the parameters